Create 1GeneralStructure.c
This program asks the user to enter her/his name and yearly income. If the income is below $30,000 the program informs the user to $0 income tax. If the income is in the range $30,000 to $100,000 the program calculates the income tax based on 15% and informs the user the amount of tax. If the income is higher than $100,000 the program calculates the tax based on 25%.